add service folder, rename files to plural
