+++ models/index.js
@@ -28,6 +28,6 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-db.Organization = require('../models/organization')(db.sequelize, Sequelize.DataTypes)
+db.Organization = require('./organizations')(db.sequelize, Sequelize.DataTypes)
 
 module.exports = db;
