+++ routes/organizations.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const organizationControllers = require('../controllers/organization')
+const organizationControllers = require('../controllers/organizations')
 
 router.get('/public', organizationControllers.getAll)
