+++ services/organizations.js
@@ -1,0 +1,10 @@
+const organizationRepo = require('../repositories/organizations')
+
+const getAll = async(req, res)=>{
+    const data = await organizationRepo.getAll()
+    return data
+}
+
+module.exports = {
+    getAll
+}
